+++ src/js/td-scenes/battle/animation/game-state/battle/attack/lightning-dozer/miss.ts
@@ -1,0 +1,21 @@
+import { Miss } from "gbraver-burst-core";
+
+import { Animate } from "../../../../../../../animation/animate";
+import { delay } from "../../../../../../../animation/delay";
+import { LightningDozerBattle } from "./lightning-dozer-battle";
+
+/**
+ * ミス
+ * @param param パラメータ
+ * @returns アニメーション
+ */
+export function miss(param: LightningDozerBattle<Miss>): Animate {
+  return param.attackerSprite
+    .charge()
+    .chain(delay(500))
+    .chain(param.attackerSprite.armHammer())
+    .chain(param.defenderSprite.avoid())
+    .chain(delay(500))
+    .chain(param.attackerSprite.hmToStand())
+    .chain(delay(500));
+}
